network warnings fixed pt2

+++ source/Utilities/Network/xlet/udp.cpp
@@ -186,7 +186,7 @@ xlet::UDPIn::UDPIn(const std::string ipstring, int port, bool qSynced) : UDPlet(
             }
             else if (n > 0)
             {
-                inDataBuffer.resize(n);
+                inDataBuffer.resize(static_cast<size_t>(n));
                 {
                     std::lock_guard<std::mutex> lock(sockMutex);
                     if (queueManaged) qin_.push_back(xlet::Data(sockAddToPeerId(cliaddr), inDataBuffer));
@@ -253,7 +253,7 @@ xlet::UDPInOut::UDPInOut(const std::string ipstring, int port, bool listen, bool
                 }
                 else if (n > 0)
                 {
-                    inDataBuffer.resize(n);
+                    inDataBuffer.resize(static_cast<size_t>(n));
                     {
                         std::lock_guard<std::mutex> lock(sockMutex);
                         if (queueManaged) qin_.push_back(xlet::Data(sockAddToPeerId(cliaddr), inDataBuffer));
